Add info in getting started docs to use matcher
I wasn't able to use smock matchers in my tests. I wasn't able to resolve my issue reading the doc, but then I noticed that in the "Basic Usage" section of the `README.md` file there is this line `chai.use(smock.matchers)`.

IMO it's a good idea to include this in the official doc as well.